Denarii GUI on Windows
The GUI now works but its hacky. We create a file to say whether we are synchronized or are synchronizing. It takes about 30 seconds to tell if denariid is stuck or not. Need to add something to the gui to indicate that we are in the process of synchronizing and mining is not possible.